refactor: 修改代码风格首页样式并提取到Program
