+++ 1st-try/src/views/Home/Program/index.vue
@@ -1,20 +1,29 @@
 <script setup lang="ts">
-import CodeRainBg from './componenets/CodeRainBg.vue'
-import FixedHeader from './componenets/FixedHeader.vue'
-import WebsiteIntroduction from './componenets/WebsiteIntroduction.vue'
-import ToolsSelector from './componenets/ToolsSelector.vue'
+import CodeRainBg from '@/views/Home/Program/components/CodeRainBg.vue'
+import HeaderMenu from '@/views/Home/publicComps/HeaderMenu/index.vue'
+import WebsiteIntroduction from '@/views/Home/Program/components/WebsiteIntroduction.vue'
+import ToolsSelector from '@/views/Home/Program/components/ToolsSelector.vue'
+
+const theme = 'dark'
+defineExpose({
+	theme
+})
 </script>
 
 <template>
   <div class="container">
     <code-rain-bg class="bg-animate" />
-    <fixed-header />
-    <div class="container-mainly">
-      <website-introduction />
-      <tools-selector class="mt-6" />
-    </div>
+    <header-menu />
+    <scrollbar>
+      <div class="container-mainly">
+        <website-introduction />
+        <tools-selector class="mt-6" />
+        <website-introduction />
+        <tools-selector class="mt-6" /><website-introduction />
+        <tools-selector class="mt-6" />
+      </div>
+    </scrollbar>
   </div>
 </template>
 
-<style scoped lang="scss" src="./program.scss">
-</style>
+<style scoped lang="scss" src="./program.scss"></style>
